Move C code into seperate files
